requirements.txt
azure-cognitiveservices-speech~=1.24.1
playsound